refactor 문자열 계산기 패키지 구조 변경
