+++ src/main/java/calculator/Number.java
@@ -1,4 +1,4 @@
-package study;
+package calculator;
 
 public class Number {
