+++ src/test/java/study/StringAddCalculatorTest.java
@@ -1,6 +1,6 @@
 package study;
 
-import org.assertj.core.api.Assertions;
+import calculator.StringAddCalculator;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EmptySource;
